requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.9.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
Django==4.2.6
django-debug-toolbar==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
flake8==6.1.0
future==0.18.3
idna==3.4
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
pep8-naming==0.13.3
Pillow==10.1.0
platformdirs==3.11.0
psycopg2-binary==2.9.9
pycodestyle==2.11.0
pyflakes==3.1.0
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rest-framework-simplejwt==0.0.2
rpds-py==0.10.6
serializers==0.2.4
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
